Make use of non local disk; adds gcp-key make target
With the current version of the chart that we use, local ssds can't
really be used given that we can't make use of PVCs for local storage,
so, in this commit we go back to using a `pd-ssd` that is big enough for
our workloads.

This commit also includes a `gcp-key` make target that grabs the gcp key
from lpass.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>