hh: add `make set-teams` to set all teams configured
On Wings, operators needed to manually set the configuration for the
teams that were part of the cluster.

By having all of the configuration in the repository, we're able to
automate that process, allowing other teams to go through the flow of
submitting a PR with what they want configured, freeing us from having
to manually do that.

This is what could be the first step to having such part automated.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>