metrics: add paging+ctx switches; expose per-cpu util in `node` dashboard
Previously we were making use of averages in the `cpu` util panel. While
that's useful to get a sense of the overall CPU utilization, it doesn't
show the real per-cpu values (as that's the nature of an `average` lol).

The effect of that was that we'd not see when a single processor (out of
16) was spinning super hard.

Here I also included context-switches and paging panels.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>