metrics: add trend to resource checks panel
resource-checks have traditionally been a good indicate of whether
workers in general are working fine, but if one looks only at the
immediate values (e.g., 30m / 1h), it's not very clear whether you're
looking at a good value or not (without going back in time).

this commit adds a dashed line that contains the value of a rolling
average over the last 24h.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>