metrics: make use of `max`es in node-pool instead of avgs
as the avgs might be deceiving when it comes to spikes and worts-case
scenarios in a per-cpu manner, this allows us to better visualize when
something is being max'ed out.

this commit also adds a separate `iowait` column as a way of letting us
better keep track of `iowait` across all nodes - something we're now
aiming at optimizing.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>