sysctl: add chart to allow us to configure worker kernel params
As reported in "Understand why system CPU usage is so high"[1], we see
that the kernel spends a lot of time trying to reclaim memory for itself
when a bunch of containers start running on those VMs.

By adjusting the minimum amount of reserved memory for the kernel to a
higher value, we're able to then let it do less work.

With the introduction of the `sysctl` chart, every time a new `worker`
node gets in, the pod in the daemonset will get scheduled to it, thus,
the parameters we want will be set on such selected set of nodes.

[1]: https://github.com/concourse/hush-house/issues/22

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>