metrics: add worker errors + pg tablestat
With the ability to get count of errors that happen on the workers,
we're able to better diagnose why workers are having trouble. Using
that, the `Concourse` dashboard now includes a `worker errors` panel to
show that.

Having `flight-recorder` constantly looking at the DB, now we also
include per-table activity and some other metrics to help us understand
our usage.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>