bump worker-ubuntu replicas + worker init fix
to get around the problem that we were facing regarding removal of state
(see https://github.com/concourse/charts/issues/6), we bumped the worker
charts to the version that contains that fix (8.2.7).

to get around the screaming SLI problem (see concourse/concourse#4553),
we bumped the web nodes (`hush-house` deployment) to use a 5.7.0 release
candidate.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>