metrics: refactor concourse dashboard
this is an incremental improvement of the readability of the `concourse`
dashboard, focusing on showing only what matters, while still making
periods configurable.

the theory behind this change is that we're usually looking to what the
"highest" of something is (rather than all of it), and in some cases, a
rate (/s) is too awkward to reason about (thus, getting rates over
longer periods of time).

ps.: both changes mentioned above are *not* hardcoded - a viewer can
change those values anytime.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>