Add k8s-topgun worker chart
This commit adds a new chart that is used to deploy a concourse worker
that connects to our ci (prod) environment. It tags itself with
`k8s-topgun` and is to be used for running the k8s-topgun test suite.

This deployment is different from other deployments in this repo. Other
deployments assume you're connected to our hush-house GKE cluster.
Instead, this deployment is meant to go on the topgun GKE cluster and
onto the node pool named `topgun-worker`.

The creds (.value.yml) file is saved in lastpass like the rest of the
deployments in this repo.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>