deployments/without-creds: remove `registry`
I had set up this release in order test out the use of container image
manifests for multiplatform images, as well as to test out how easy to
have a non-harbor/dockerhub registry would look like, but never remove
it.

as we make no use of this anywhere, it's time to remove it.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>