ci: enable Lets Encrypt
concourse/concourse#5154

In order to do this, we had to first ship a new version of the chart,
because there was a bug in the Lets Encrypt logic:
concourse/concourse#5284

It turned out that, in order to deploy this new version of the chart,
we actually needed to start using Helm v3 instead of Helm v2:

When first trying to enable Lets Encrypt, we uncovered the
above-mentioned bug - the web pods were stuck in a crash loop. However,
when we tried to apply a fix, the tiller didn't make any changes - it
was as though it didn't "see" the diff. @cirocosta said it would be wise
to use
https://helm.sh/docs/faq/#improved-upgrade-strategy-3-way-strategic-merge-patches
to resolve the issue.

We're now committed to using Helm v3 to manage the `ci` release, but we
don't want to force this decision on the other releases just yet.
Accordingly, we split out a separate Makefile for `ci` which uses Helm v3
conventions. However, the old Makefile now delegates to the new one, so
the usual workflows should still work.

Signed-off-by: Rui Yang <ryang@pivotal.io>
Co-authored-by: Jamie Klassen <cklassen@pivotal.io>