Add app behind hangar.concourse-ci.org
This is an nginx server that does a redirect to a zoom room used by the
team for meetings like: standup, retro, everyone is wfh because covid-19

You may wonder: why do the redirect with an html page? You can do HTTP
301 redirect with just nginx redirect rules.

and you would be correct! But then other people's opinions get pushed
onto you and you have to do hacky things to get around it.

tldr: the ingress comes with a health checker that is not smart. See
https://github.com/kubernetes/ingress-gce/issues/42 for more details.

Longer answer: The ingress used by GKE setups a health check that ONLY
hits '/' against your pod. Because my initial nginx.conf just did 301
redirects this healthcheck was forever failing because it would never
get HTTP 200. I tried to get around this by having nginx listen on
another port for health checking and always return HTTP 200 on that
port. I then had to expose both ports to the ingress controller. The
ingress controller decided to healthcheck BOTH ports that were exposed
on the container and send all traffic to port 9000 where it was getting
the HTTP 200's. At this point I threw my hands up and wrote some HTML to
do the redirect for me.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>