Add elasticsearch/kibana to jaeger deployment
concourse/concourse#5382
fixes concourse/hush-house#119

also notable that we are now using helm3 for this.

Along the way, we found that for some deploys the number of replicas for the
jaeger-related indices in elasticsearch was being bumped to 1, which caused
the cluster status to go yellow, since we are only using a single node.

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>
Co-authored-by: Ciro S. Costa <cscosta@pivotal.io>