ci/hush-house: use 6.1 release
as we wanted to get the 6.1 release candidate before shipping it, we
went ahead with using a `file://` reference that points to a local
version of `concourse/concourse-chart` (essentially, `release/6.1.x`).

this can be updated to a reference in the bucket once we get `6.1.x`
release from a `concourse/concourse-chart` perspective.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>