send metrics to both wavefront + datadog
by serving a prometheus endpoint, which will be scraped by two agents:

* opencensus, which will forward to wavefront; and
* telegraf, which will forward to datadog.

I haven't tested whether the telegraf plugin for datadog output plays well
with a unix domain socket. There are other ways to configure it to send metrics
directly to the datadog API and bypassing the agent socket if necessary.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>