export traces to wavefront directly
The oc-agent is outdated and it's recommended we switch to the
opentelementry-collector. However, since Wavefront is capable of
ingesting traces in Jaeger format, let's first try to get the traces to
show up in wavefront before adding any more machinery.

Along the way I discovered that wavefront rejects any spans that have
more than 20 tags, the tag value has to be less than 255 characters, and
other fun stuff. I've put a bandaid over it by dropping some tags that
were inserted by the library we use, but this can come back to bite us
in the future.

Signed-off-by: Bohan Chen <bochen@pivotal.io>