switch metrics from datadog to wavefront
- the otel-collector now scrapes the web for prometheus metrics and
exports it to a prometheus storage adapter
- the prom-storage-adapter is an integration provided by wavefront to
ingest metrics from open-telemetry and output to the proxy:
https://docs.wavefront.com/opentelemetry.html
- switchied to otel-collector-contrib so that we can get access to the
`metricstransform` processor so that we can add tags as a preprocessor
step
- we need to add `url` tag so that the wavefront metrics dashboard
template can differentiate between different clusters
- turns out there's a bug where any metrics `Attributes` configured on
concourse will get ignored by the prometheus emitter. So we're forced to
add tags using the otel-collector (which is also a neat demonstration of
what it can do)
- combined the tracing and built-in concourse ci metrics dashboard from
wavefront into a single one. Also grabbed the json format in case we
need it again
- kept the datadog sub-chart because hush-house is currently relying on
that

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Bohan Chen <bochen@pivotal.io>