restrict wavefront-proxy's memory usage
The container kept growing in memory usage until it brought down the
entire pod via OOM killer

We tried to set it up as a standalone deployment, but we ran into
problems of figuring out which host/pod a metric/trace came from. We
tried to solve that with the otelcol-contrib's k8sprocessor, but that
required the ci-web service account to have special rolebindings that we
couldn't get working.

So instead we just limited the memory of the container and 500Mb seems
like a decent number, let's see if it gets overloaded anytime soon.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Bohan Chen <bochen@pivotal.io>