Add resources requests and limits to all sidecars
Initially the QOS class of the ci-web pods were "guaranteed". The QOS
class dropped to "burstable" when sidecars were added that didn't have
resources defined. This resulted in the containers getting OOM killed.

It's still unclear why the node needed to OOM kill the containers as it
was not under any memory or CPU pressure based on the metrics we have.

The image sha drifted again. I updated it to what was currently
deployed to avoid any accidental upgrade/downgrade issue.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>